修改

+++ src/main/java/com/sevencolor/domain/dao/RebalanceStatisticsInfoMapper.java
@@ -3,13 +3,13 @@ package com.sevencolor.domain.dao;
 import com.sevencolor.domain.pojo.RebalanceStatisticsInfo;
 
 public interface RebalanceStatisticsInfoMapper {
-  int deleteByPrimaryKey(Integer dbid);
+  int deleteByPrimaryKey(Long dbid);
 
   int insert(RebalanceStatisticsInfo record);
 
   int insertSelective(RebalanceStatisticsInfo record);
 
-  RebalanceStatisticsInfo selectByPrimaryKey(Integer dbid);
+  RebalanceStatisticsInfo selectByPrimaryKey(Long dbid);
 
   int updateByPrimaryKeySelective(RebalanceStatisticsInfo record);
